Conflicts:
	eclipse_ws/biobank.common/src/edu/ualberta/med/biobank/common/wrappers/AbstractObjectWithPositionManagement.java
	eclipse_ws/biobank.common/src/edu/ualberta/med/biobank/common/wrappers/ContainerWrapper.java
	eclipse_ws/biobank.common/src/edu/ualberta/med/biobank/common/wrappers/SiteWrapper.java
	eclipse_ws/biobank.common/src/edu/ualberta/med/biobank/common/wrappers/SpecimenWrapper.java
	eclipse_ws/biobank2.tests/src/edu/ualberta/med/biobank/test/internal/SpecimenHelper.java
	eclipse_ws/biobank2.tests/src/edu/ualberta/med/biobank/test/wrappers/TestProcessingEvent.java
	eclipse_ws/biobank2.tests/src/edu/ualberta/med/biobank/test/wrappers/TestSite.java
	eclipse_ws/biobank2.tests/src/edu/ualberta/med/biobank/test/wrappers/TestSpecimen.java
	eclipse_ws/biobank2/src/edu/ualberta/med/biobank/forms/linkassign/AbstractLinkAssignEntryForm.java
	eclipse_ws/biobank2/src/edu/ualberta/med/biobank/forms/linkassign/SpecimenAssignEntryForm.java

+++ eclipse_ws/biobank2.tests/src/edu/ualberta/med/biobank/test/internal/CenterHelper.java
@@ -15,7 +15,8 @@ public class CenterHelper extends DbHelper {
         for (SpecimenWrapper spc : center.getSpecimenCollection(false)) {
             if (spc.getOriginInfo().getCenter().equals(center)
                 && (spc.getParentSpecimen() != null)) {
-                spc.delete();
+                if (!spc.isNew())
+                    spc.delete();
             }
         }
 
@@ -28,7 +29,8 @@ public class CenterHelper extends DbHelper {
                 if (shipInfo != null) {
                     shipInfo.delete();
                 }
-                spc.delete();
+                if (!spc.isNew())
+                    spc.delete();
             }
         }
         deleteFromList(center.getOriginInfoCollection(false));
